Changes for vectors (from int to double)

+++ Clustering/include/Helper_Functions.h
@@ -6,7 +6,7 @@ template <typename Point> double min_distance(int, vector<int>*, vector<vector<P
 
 void show_cluster_usage(string name);
 int Read_input_file(string input);
-int Read_files(vector<vector<int>>*, int*, string, string );
+int Read_files(vector<vector<double>>*, int*, string, string );
 int Read_files(vector<vector<double*>>*, int*, string, string);
 double point_dist(double*, double*, int=2);
 double DTW(vector<double*>*, vector<double*>*);
